cron DriverActiveState:FirstReminder based on service provider id

+++ app/Console/Commands/MakeUnactiveDriverLogout.php
@@ -43,19 +43,23 @@ class MakeUnactiveDriverLogout extends Command
      */
     public function handle()
     {
-        $settings = Setting::first();
-        $settingValue = json_decode($settings['value']);
-        $alertTime = Carbon::now()->subSeconds($settingValue->waiting_time)->format('Y-m-d H:i:s');
-        $unactive_driver_ids = DriverStayActiveNotification::where('last_activity_time', '<=', $alertTime)->where(['is_availability_alert_sent' => 1, 'is_availability_changed' => 1, 'is_logout_alert_sent' => 1])->pluck('driver_id')->toArray();
-        if (count($unactive_driver_ids) > 0) {
-            DB::table('oauth_access_tokens')
-                ->whereIn('user_id', $unactive_driver_ids)
-                ->update([
-                    'revoked' => true
-                ]);
-            User::whereIn('id', $unactive_driver_ids)->update(['availability' => 0, 'fcm_token' => '', 'device_token' => '']);
-            DriverStayActiveNotification::whereIn('driver_id', $unactive_driver_ids)->delete();
-            DriverChooseCar::whereIn('user_id', $unactive_driver_ids)->where(['logout' => 0])->update(['logout' => 1]);
+        $overallSettings = Setting::get();
+        if (!empty($overallSettings) && count($overallSettings) > 0) {
+            foreach ($overallSettings as $setting) {
+                $settingValue = json_decode($setting['value']);
+                $alertTime = Carbon::now()->subSeconds($settingValue->waiting_time)->format('Y-m-d H:i:s');
+                $unactive_driver_ids = DriverStayActiveNotification::where('last_activity_time', '<=', $alertTime)->where(['is_availability_alert_sent' => 1, 'is_availability_changed' => 1, 'is_logout_alert_sent' => 1, 'service_provider_id' => $setting->service_provider_id])->pluck('driver_id')->toArray();
+                if (count($unactive_driver_ids) > 0) {
+                    DB::table('oauth_access_tokens')
+                        ->whereIn('user_id', $unactive_driver_ids)
+                        ->update([
+                            'revoked' => true
+                        ]);
+                    User::whereIn('id', $unactive_driver_ids)->update(['availability' => 0, 'fcm_token' => '', 'device_token' => '']);
+                    DriverStayActiveNotification::whereIn('driver_id', $unactive_driver_ids)->delete();
+                    DriverChooseCar::whereIn('user_id', $unactive_driver_ids)->where(['logout' => 0])->update(['logout' => 1]);
+                }
+            }
         }
     }
 }
